Added a function to change desktop names "On the run".
Feature proposed by @achim-t in #60.

Added a new customizable shortcut (and updated documnetation accordingly)
to show a popup which lets the user dinamically change the name of the current desktop.
The change is not permanent and will be overwritten by the default values in
'settings.ini' when the program will be reloaded.