Fixed focus management.
Rewrote the `_Focus()` function to give focus to the foremost window in the desktop.

Replaced the `_FocusIfRequested()` function with `_Focus()`.

Had to update the library to the latest version as the `IsWindowOnDesktopNumber` function seemed not to work in the version previoulsy used.

This solves #64 and possibly even the glitch of the flashing orange icons in the taskbar: I could not reproduce it in my tests for this version of the code, while with the stable release it happened quite often.