Created script to automate the deployment of anew version.
The 'deploy-release/deploy.py' is a Python script that automates the steps required to deploy a new release.
After some security checks it copies the necessary files, discarding those which are not needed, in a remp directory, converts .md files to .html fixing links as well, then cretes the zip of the new release.